polars_docker/requirements.txt
polars
python-dotenv
requests
